博客相关修改

+++ blog/src/main/java/com/wcf/funny/blog/entity/ArticleInfo.java
@@ -62,5 +62,5 @@ public class ArticleInfo extends ArticleSimple implements Serializable {
      *
      * @return
      */
-    private Short deleteFlag;
+    private Integer deleteFlag;
 }
